Re-added selection tests
